move compat to separate file
