+++ Pod/Classes/URBNLengthRules.swift
@@ -95,17 +95,3 @@ public class URBNMaxLengthRule: BaseLengthRule {
     
     override var comparisonFunc: (lhs: Int, rhs: Int) -> Bool { return isInclusive ? (<=) : (<) }
 }
-
-@objc public class CompatMinLR: CompatBaseRule {
-    public init(minLength: Int, inclusive: Bool = false, localizationKey: String? = nil) {
-        super.init()
-        self.baseRule = URBNMinLengthRule(minLength: minLength, inclusive: inclusive, localizationKey: localizationKey)
-    }
-}
-
-@objc public class CompatMaxLR: CompatBaseRule {
-    public init(maxLength: Int, inclusive: Bool = false, localizationKey: String? = nil) {
-        super.init()
-        self.baseRule = URBNMaxLengthRule(maxLength: maxLength, inclusive: inclusive, localizationKey: localizationKey)
-    }
-}
