+++ Pod/Classes/URBNValidationRules.swift
@@ -8,7 +8,6 @@
 
 import Foundation
 
-
 public protocol ValidationRule {
     var localizationKey: String { get set }
     func validateValue(value: Any?) -> Bool
@@ -18,7 +17,6 @@ public protocol ValidationRule {
 public protocol URBNRequirement {
     var isRequired: Bool { get set }
 }
-
 
 public class URBNBaseRule: ValidationRule {
     var _localizationKey: String?
@@ -92,34 +90,6 @@ public class URBNBlockRule: URBNBaseRule {
     }
 }
 
-@objc public class CompatBlockRule: CompatBaseRule {
-    public typealias CompatBlockValidation = (value: AnyObject?) -> Bool
-
-    public init(_ validator: BlockValidation) {
-        super.init()
-        let blockRule = URBNBlockRule(validator)
-        blockRule.blockValidation = validator
-        self.baseRule = blockRule
-    }
-    public init(validator: BlockValidation, localizationKey: String? = nil) {
-        super.init(localizationKey: localizationKey)
-        let blockRule = URBNBlockRule(validator: validator, localizationKey: localizationKey)
-        blockRule.blockValidation = validator
-        self.baseRule = blockRule
-    }
-    public convenience init(validator: CompatBlockValidation, localizationKey: String?) {
-        let valB = { (v: Any?) -> Bool in
-            if let anyObjectv = v as? AnyObject {
-                return validator(value: anyObjectv)
-            }
-            
-            return false
-        }
-        
-        self.init(validator: valB, localizationKey: localizationKey)
-    }
-}
-
 public class URBNRegexRule: URBNBaseRule, URBNRequirement {
     internal var pattern: String
     public var isRequired: Bool = false
@@ -134,69 +104,5 @@ public class URBNRegexRule: URBNBaseRule, URBNRequirement {
         
         let pred = NSPredicate(format: "SELF MATCHES[cd] %@", self.pattern)
         return pred.evaluateWithObject(value as! AnyObject?)
-    }
-}
-
-@objc public protocol OCValidationRule {
-    var localizationKey: String { get set }
-    func validateValue(value: AnyObject?) -> Bool
-    func validateValue(value: AnyObject?, key: String) -> Bool
-}
-
-@objc public class CompatBaseRule: NSObject, OCValidationRule {
-    var baseRule: URBNBaseRule
-    
-    public init(localizationKey: String? = nil) {
-        baseRule = URBNBaseRule(localizationKey: localizationKey)
-        super.init()
-    }
-    
-    public func validateValue(value: AnyObject?) -> Bool { return baseRule.validateValue(value) }
-    public func validateValue(value: AnyObject?, key: String) -> Bool { return baseRule.validateValue(value, key: key) }
-    public var localizationKey: String  {
-        get {
-            return baseRule.localizationKey
-        }
-        set {
-            baseRule.localizationKey = newValue
-        }
-    }
-}
-
-@objc public class CompatRegexRule: CompatBaseRule, URBNRequirement {
-    public var isRequired: Bool {
-        get {
-            if let requiredBaseRule = self.baseRule as? URBNRegexRule {
-                return requiredBaseRule.isRequired
-            }
-        
-            return false
-        }
-        set {
-            if let requiredBaseRule = self.baseRule as? URBNRegexRule {
-                requiredBaseRule.isRequired = newValue
-            }
-        }
-    }
-    
-    public init(pattern: String, localizationKey: String? = nil) {
-        super.init()
-        self.baseRule = URBNRegexRule(pattern: pattern, localizationKey: localizationKey)
-    }
-    
-    public static let emailPattern = "^(?:(?:(?:(?:[a-zA-Z0-9_!#\\$\\%&'*+/=?\\^`{}~|\\-]+)(?:\\.(?:[a-zA-Z0-9_!#\\$\\%&'*+/=?\\^`{}~|\\-]+))*)|(?:\"(?:\\\\[^\\r\\n]|[^\\\\\"])*\")))\\@(?:(?:(?:(?:[a-zA-Z0-9_!#\\$\\%&'*+/=?\\^`{}~|\\-]+)(?:\\.(?:[a-zA-Z0-9_!#\\$%&'*+/=?\\^`{}~|\\-]+))*)|(?:\\[(?:\\\\\\S|[\\x21-\\x5a\\x5e-\\x7e])*\\])))$"
-}
-
-@objc public class CompatRequiredRule: CompatBaseRule {
-    public override init(localizationKey: String? = nil) {
-        super.init(localizationKey: localizationKey)
-        self.baseRule = URBNRequiredRule()
-    }
-}
-
-@objc public class CompatNotRequiredRule: CompatBaseRule {
-    public override init(localizationKey: String? = nil) {
-        super.init(localizationKey: localizationKey)
-        self.baseRule = URBNNotRequiredRule()
     }
 }
